Rewrite move generation.
Inspired from Sunfish. As it is, checks for own king in check
after move is not done. Castling is currently possible through a enemy
ray piece, fixed by the use of king enpassant.

The thinking is that moves that leave the king open for capture *could*
be generated, but they will fail quickly because the king is lost on
next move. So a searcher should never end up giving a move that is
illegal. At least that is the current justification.