Add first edition of search.
Simple alpha beta pruning, with some very crude
implementation of pv construction.

Next steps should be implementing Zobrist hashing
so that a transposition table can be used. Then an
implementation of MTD-bi search algorithm as used
in Sunfish. The hope is that that algorithm (including
some of the pruning/move ordering used there) provides
a clean and good functioning engine, which could then be
cleaned up and maybe enter a somewhat finished state.