Add LRUCache implementation.
Only very slightly modified from original. When inserting
a key value pair, we use a copy and not a reference. Reason
being that the the key/values are created somewhere in the
search tree, and the references may expire before they are
requested. This issue has not been directly observed, but
the change remains as it gives some peace of mind.