Fix castling in when playing with UCI.
Castle moves were not encoded properly before sent to doMove,
and so castling was not handled right.

The engine is still terrible, and the reported scores seem to very
dramatically from turn to turn. Possibly due to a bug in the search, or
just an artifact of it being so stupid.