Add match running script.
Does not add anything more than what cutechess-cli is, but this way it
will be a bit easier to change settings, rerun same matches, and
generally remember how to use cutechess-cli.

The following is the result of running the current version against the
version which is equivalent to the original fluxroot/pulse engine. The
difference is only the addition of null move pruning.

Time control was 40/60 (40 moves in 60 seconds). Note that the ratings
only have meaning relative to each other, as the absolute value is not
chored to any known ELO rating.

    # PLAYER          :  RATING  POINTS  PLAYED   (%)
    1 goldfish-new    :  2327.6    18.5      32    58
    2 goldfish-old    :  2272.4    13.5      32    42