Refactor code base (#15)
* Major refactor towards use of enums and scope changes.

As of now though, there is WAY to many static_cast. Did this first,
so that everything is explicit. Can now go to implicit conversions,
knowing there aren't any stray integer representations left.

* Remove all unnecessary static_cast.

Enums are now not scoped anymore to allow for implicit int conversion.
This is a bit of a bummer, but I will make a point to use them as
scoped, and hopefully all (or most) unscoped usages have been removed. I
still think this is worth  it, as the static_cast's got to be a bit
much.