Implement aspiration window.
Seems to be good. One issue is with benchmark position 38, where the
mate in 5 is not found. The best move is still found, but its not
finding the mate. For some reason the search returns a value in [alpha,
beta] even when beta << CHECKMATE_THRESHOLD.

Crossing fingers and hoping that this doesn't mean anything bad.