Stop searching deeper when optimal mate is found.
If we have searched to a depth d, and evaluated the position as a mate
in <= d plies (for either side), then nothing at a deeper level will
change that evaluation. So we can safely stop the search. Improves time
to move in such cases (not vital, but a reasonable addition).

Added as a step in the direction of solving #25.