Bugfix: Be more restrictive about using NMP
Null-move pruning was previously restricted to be used when beta <= CHECKMATE,
which essentially was no limit at all. We only do NMP for a potential
beta-cutoff, i.e. the null-move-score is >= beta. But we also don't
trust null-move search when it gives mate scores. These two things don't
match up.

So, we only use NMP when beta is less than any mating score.

Also added is the limitation that we do not do two successive
null-moves. This just amounts to doing a reduced search of the initial
position, and won't be much help. It's currently still allowed with more
than one null-move in total in a variation though.

This is potentially a fix for #25. However, this will not be closed before more
rigorous testing has been done.