Bugfix: Consistent retrieval of mates form tt
Was not being careful with the difference between how mates are
interpreted when they are stored vs. when they are retrieved (plies to
mate from the root vs. plies to mate from the current position). Passing
the value through the lightweight functions added here ensures that this
is handled consistently.

This issue resulted in bugs where Goldfish reported mates which where
shorter than possible. This is another candidate for the bugs seen in #25,
which now might be resolved (to be verified).