Perform move sorting/picking in probe_root
Calling probe_root now deletes all the moves that don't preserve the
win/draw/loss, and so regular search on the remaining moves cannot go
wrong. This is better than just picking the top move from tb, because
DTZ is not always desired to minimize. We want to minimize DTM, but just
making sure that we do so without losing the position.