Use root move probing to play optimal endgames
If the root position is in TB, Goldfish should now play optimal, in the
sense that it will obtain the best possible score from the position. It
is not guaranteed to play the fastest mating sequence however, but will
choose from the possible best moves using regular search after tb
filtering.