Respect contempt value in search
Not used at root (no search started in the case of stalemate at root, so
no need), but respected everywhere else.

A large contempt value will result in GF treating a draw as a good
result (i.e. it has respect for it's opponent and will take a draw if it
can). A large negative value will cause GF to avoid draws as best it
can.