Implement crude killer move heuristic
Any move that is determined as best at a given ply is stored as a killer, and
killers are given a tie-break in move sorting when occurring in sibling
nodes. Although this is normally only used to sort quiet moves, since we
have no notion of good/bad captures (i.e. no SEE), this works better
when also allowing this to improve ordering for captures as well.

Sees a solid ~20% decrease in bench nodes. Because such a simple
addition yields such a large bonus, this indicates that there is quite a
lot of gain to be had w.r.t move ordering. Should likely spend more time
on this going forward.