Fix correctness bug in quiescent search
Bug also present in 1.x version of Goldfish.

In short, in positions where the side to move is in check
and there are no legal captures, quiescent search would return
a value of -INFINITE because it filters out all the non-capturing
evasions and never updates the default starting value of -INFINITE.

In positions where the side to move is in check we should search all
legal moves, i.e. all legal captues + evasions.