Fix problem with infinite PVs
The chess programming wiki says to use only the exact TT entries for PV, which makes sense.
This alone fixes the problem in the cases where it has been seen to loop indefinetly.
An example is R7/P4k2/8/8/8/8/r7/6K1 w - - 0 1, which hits an infinite loop quickly.

As a safeguard/optimization, we also limit the PV to at most 8 ply. Don't need more than that
for verification purposes, and this will primarily play without direct supervision anyway.

Might consider lowering this further if it proves to be a bottleneck. Not likely though.