Setup search suite benchmarks with criterion
Currently the lack of search pruning makes the search so slow that many
of the existing positions can't be searched to depth 6 as with v1.13.

Will add more positions as we go, but even just the basic starting
position is enough of a test case to establish some early gains.