Change to using the new uci API
This involves running the search in the main thread, instead of dispatching a thread.
This looks and feels much better, and allows removing some arcs and some mutexes.

Also now using UCI options in place of hard-coded values all over.
This is a little less performant, sure, but allows easier tuning.
Generally a win, I don't think the additional memory reads will be the bottleneck.