Fix killer heuristic
This was broken badly, leading to very bad move ordering. The killers
are now only added if they are quiet moves that caused a beta cutoff, as
they are intended to be. The influence this has on move ordering is also
fixed. They are now considered right after tt move and winning captures,
not just after tt.

Causes a drastic improvement in the performance of the engine.