Fix draw-by-repetition detection
It would not count pre-root moves in the detection, meaning it only
found repetitions in-search. This should hopefully cause a _lot_ less
draw-by-repetition in play tests, and less/no drawing in won positions.

Edit: Still a lot of draw-by-repetition in play tests. But spot-checking
them, they all make sense. Positions are entirely equal, and the side
with slightly less control semi-forces the repetition with checks. Time
will show if this still draws won positions. We probably get a lot of
repetition because the evaluation is so simple, and nobody knows how to
make progress in early endgames.